feat: move HTMLCanvas class to presentation layer
