+++ src/components/Canvas/CanvasFooter.vue
@@ -81,4 +81,3 @@ export default defineComponent({
   },
 })
 </script>
-@/application/services/interpolator
