+++ src/components/Canvas/CanvasHeader.vue
@@ -62,4 +62,3 @@ export default defineComponent({
   },
 })
 </script>
-@/application/services/interpolator
