+++ src/components/Canvas/CanvasPlot.vue
@@ -159,4 +159,3 @@ export default defineComponent({
   },
 })
 </script>
-@/application/services/interpolator
