+++ src/components/Magnifier/MagnifierPlots.vue
@@ -144,4 +144,3 @@ export default defineComponent({
   },
 })
 </script>
-@/application/services/interpolator
